feat: add enrollment_date and custom fields to profile data csv
Add `enrollment_date` column on the csv file of all students enrolled
in a course.

Allow site operators to include on the export of profile information as CSV
custom fields if the platform has an extending User model.
This can be used if you have an extended model that include for example
an university student number and site operator want to export the
student number on the student profile information CSV.

GN-914